Android Rally
It is an Android pplication that is basically a location based game.In
this game the user basically has two modes to play with the the Rally
one is automatic mode and the other is Custom mode. For  now this
application is single user but I am developing it to be a multi-user
game.